unet_pretraining/requirements.txt
numpy
matplotlib
Pillow
opencv
scikit-learn
pytorch
torchvision
pytorch-cuda=11.8
monai
